moved dynamic manifest generation to App.tsx
