+++ src/containers/App.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import { Route, Switch, Router, useLocation } from 'react-router-dom'
 import analytics from 'universal-ga'
 
@@ -49,6 +49,71 @@ function getDashboardComponent(
     }
 }
 
+function updateManifest(pathName: string): void {
+    const manifest = window.document.getElementById('manifest-placeholder')
+    if (manifest) {
+        const dynamicManifest = {
+            name: 'Tavla - Enturs avgangstavle',
+            short_name: 'Tavla',
+            start_url: `.${pathName}`,
+            display: 'standalone',
+            background_color: '#181C56',
+            theme_color: '#181C56',
+            description: 'Lag din egen sanntidstavle.',
+            orientation: 'portrait',
+            lang: 'no',
+            images: [
+                {
+                    src: 'images/logo/logo-72x72.png',
+                    sizes: '72x72',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-96x96.png',
+                    sizes: '96x96',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-128x128.png',
+                    sizes: '128x128',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-144x144.png',
+                    sizes: '144x144',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-152x152.png',
+                    sizes: '152x152',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-192x192.png',
+                    sizes: '192x192',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-384x384.png',
+                    sizes: '384x384',
+                    type: 'image/png',
+                },
+                {
+                    src: 'images/logo/logo-512x512.png',
+                    sizes: '512x512',
+                    type: 'image/png',
+                },
+            ],
+        }
+        const stringManifest = JSON.stringify(dynamicManifest)
+        const blob = new Blob([stringManifest], {
+            type: 'application/json',
+        })
+        const manifestURL = URL.createObjectURL(blob)
+        manifest.setAttribute('href', manifestURL)
+    }
+}
+
 const Content = (): JSX.Element => {
     const user = useFirebaseAuthentication()
     const settings = useSettings()
@@ -59,6 +124,10 @@ const Content = (): JSX.Element => {
     const Dashboard = settings[0]
         ? getDashboardComponent(settings[0].dashboard)
         : (): null => null
+
+    useEffect(() => {
+        updateManifest(location.pathname)
+    }, [location.pathname])
 
     return (
         <UserProvider value={user}>
